+++ bot/general/values.py
@@ -60,8 +60,6 @@ USER_EMOJI = u"\U0001F464"
 
 ADMIN_LIST = [247340182]
 
-BOT_USERNAME = "GamersUtilityBot"
-BOT_URL = f"https://t.me/{BOT_USERNAME}"
 TOKEN = os.environ.get("GUB_BOT_TOKEN")
 XBOX_CLIENT_SECRET_EXPIRY_DATE = datetime.datetime.strptime("Jun 16, 2024", "%b %d, %Y").date()