+++ bot/handlers/notify_groups/modification.py
@@ -93,9 +93,12 @@ def start(update, context):
                 strings.BOT_UNABLE_TO_SEND_PRIVATE_MESSAGE(
                     context.user_data["user_mention"],
                     "modifying a notify group",
-                    "modify\_notify\_group"
+                    "/modify\_notify\_group"
                 ),
-                reply_markup=inline_keyboards.go_to_private_chat_keyboard(),
+                reply_markup=inline_keyboards.go_to_private_chat_keyboard(
+                    context.bot.get_me().username,
+                    f"mng_{context.user_data['chat_id']}"
+                ),
                 parse_mode=ParseMode.MARKDOWN_V2
             )
         else: