+++ bot/handlers/status/add_status_user.py
@@ -70,9 +70,12 @@ def start(update, context):
                 strings.BOT_UNABLE_TO_SEND_PRIVATE_MESSAGE(
                     user_mention,
                     "adding a status user",
-                    "add\_status\_user"
+                    "/add\_status\_user"
                 ),
-                reply_markup=inline_keyboards.go_to_private_chat_keyboard(),
+                reply_markup=inline_keyboards.go_to_private_chat_keyboard(
+                    context.bot.get_me().username,
+                    f"asu_{context.user_data['chat_id']}"
+                ),
                 parse_mode=ParseMode.MARKDOWN_V2
             )
         else: