+++ CHANGELOG.md
@@ -15,6 +15,8 @@ command arguments to the /list_notify_groups command.
 /list_notify_groups command.
 ### Changed
 - Fixes the v0.1.0 link for CHANGELOG.md
+- Instead of setting the notify group names as pre-formatted, the bot will bold
+them when a notify group is to be stringized.
 ### Removed